improves files tree (join folders ui and uis and move `ui/utils/utils.tsx` to `utils/ui-utils.tsx`)
