+++ scene/src/ui/creation-player/creationPlayer.tsx
@@ -1,6 +1,6 @@
 import { UiCanvasInformation, engine } from '@dcl/sdk/ecs'
 import ReactEcs, { UiEntity } from '@dcl/sdk/react-ecs'
-import { type Sprite, getUvs } from '../utils/utils'
+import { type Sprite, getUvs } from '../../utils/ui-utils'
 import {
   CHARACTER_CLASSES,
   CHARACTER_FACTIONS,
