+++ scene/src/ui/magical-market/magicalMarket.tsx
@@ -1,6 +1,6 @@
 import { UiCanvasInformation, engine } from '@dcl/sdk/ecs'
 import ReactEcs, { UiEntity } from '@dcl/sdk/react-ecs'
-import { getUvs } from '../utils/utils'
+import { getUvs } from '../../utils/ui-utils'
 import type { MagicalItemsMarketProp } from './magicalMarketData'
 import {
   HEIGTH_FACTOR,
